# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "tinrist123@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tags/post/{id}/page={page}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Get post by id tag and show on per page",
                "description": "Returns post related with id tag",
                "operationId": "getPostByIdTagAndPage11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tag id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "page  number ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/tags/post/total/{id}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Get total post by id tag ",
                "description": "Returns total post related with id tag",
                "operationId": "getTotalPostByIdTag11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tag id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/tags/post/{id}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Get  post by id tag ",
                "description": "Returns  post related with id tag",
                "operationId": "getPostByIdTag1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tag id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/tag": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Get list of tag",
                "description": "Returns list of tag",
                "operationId": "getAllTag",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}